Create README.md
I graduated from Istanbul Sabahattin Zaim University. College of Computer and Software Engineering. Specialized in web design, web pages, and also specialized in programming and developing systems and websites through Back-end : php - oop - mysql - laravel.
I worked in this field with two companies. The first company in Turkey I worked with for 3 years is a software development company. The second is a Malaysian company also based in Turkey, and I worked with it for a year and a half.
Also, I have taught Front-end:  html - css - javascript - jquery - bootstrap, and I am also a specialist in it, as I worked in it for four years. I also worked as a freelancer programmer and have many projects that I developed and programmed.